Add post types (#77)
Adds a drop down to choose the post type you want to create (e.g. a Page). This also changes the category and tag inputs to hide on posts that don't support them.

Note that for a post type to appear in the list, the value from /post-types needs to have `supports.editor` - this keeps "Media" off the list, and any custom post types that don't have a place for our HTML.

We also look to see if we have saved the postTypes. If not, it will update the site - presumably because the user is migrating from the old version.